pyproject.toml
[project]
name = "calendai-pro"
version = "0.1.0"
description = "AI-powered scheduling web application with Flask, LangChain, Supabase, and Google Calendar sync"
requires-python = ">=3.11"
dependencies = [
    # Flask core
    "Flask>=3.0,<4.0",
    "Flask-Login>=0.6,<1.0",
    "Flask-SQLAlchemy>=3.1,<4.0",
    "Flask-WTF>=1.2,<2.0",
    "gunicorn>=22.0,<23.0",

    # Database
    "SQLAlchemy>=2.0,<3.0",
    "alembic>=1.13,<2.0",
    "psycopg2-binary>=2.9,<3.0",

    # Auth
    "Authlib>=1.3,<2.0",

    # AI / LangChain
    "langchain>=0.3,<1.0",
    "langchain-openai>=0.3,<1.0",
    "langchain-core>=0.3,<1.0",

    # Schemas & Validation
    "pydantic>=2.0,<3.0",

    # Date/Time
    "python-dateutil>=2.9,<3.0",
    "pytz>=2024.1",
    "tzdata>=2024.1",

    # Security
    "cryptography>=43.0,<44.0",

    # HTTP client (for Google Calendar API)
    "httpx>=0.27,<1.0",
    "google-api-python-client>=2.0,<3.0",
    "google-auth>=2.0,<3.0",
    "google-auth-httplib2>=0.2,<1.0",

    # Utils
    "python-dotenv>=1.0,<2.0",
    "APScheduler>=3.10,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=5.0,<6.0",
    "pytest-flask>=1.3,<2.0",
    "factory-boy>=3.3,<4.0",
    "faker>=28.0,<29.0",
    "ruff>=0.6,<1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
